Update
Added option to have a sort of "SecretPhrase" to bypass asking the user to run non-admin commands via TCP. The secretphrase file should be stored on a seperate device or server in order to prevent accidental removal or edits by the end user.

Added popup confirming is the end user was to accept non-standard TCP commands, ie: things not specifically allowed to run via TCP like updates or restarting the POS.

Added ability to run TCP commands as admin if the string $Admin is present in the TCP string.

I think this was added in the last commit but whatever,
Added checks to prevent the button from being pulled too far off of the screen.